% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fbread}
\alias{fbread}
\title{A simplified read batches function using fread().}
\usage{
fbread(path = "./data", pattern = NULL, ...)
}
\arguments{
\item{path}{char. The files' path to be loaded. The default corresponds to the the './data' under working directory.}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be returned.}

\item{...}{arguments in fread(). Details refer to ?fread().}
}
\description{
A simplified read batches function using fread().
}
\examples{
fbread(pattern = "*.csv")
fbread(pattern = "*.csv", encoding = "UTF-8")
}
