% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillna.r
\name{fillna}
\alias{fillna}
\title{Fill NAs with user-specified value}
\usage{
fillna(x, na.value = 0, fill.inf = F, inf.value = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{na.value}{Value for NAs}

\item{inf.value}{Value for Inf or -Inf}
}
\description{
Fill NAs with user-specified value
}
