% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.r
\name{growth}
\alias{growth}
\title{Calculate Decimal Growth Rate.}
\usage{
growth(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The decimal growth rate of vector \code{x}. The first element will be NA.
}
\description{
\code{growth} return the decimal growth rate of a numeric vector. It employs data.table:::shift() to achieve the best performance.
}
\examples{
growth(c(1, 2, 1))
}
