% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sv}
\alias{sv}
\title{A simplified save function using saveRDS() or save().}
\usage{
sv(objname, filename = NULL, path = "./data", compress = T, rds = T)
}
\arguments{
\item{objname}{expression or char. The object to be saved, use its R object name, e.g., sv(dt).}

\item{filename}{expression or char. The file name of the serialized object. Default to objname, e.g., sv(dt, dt_new_name)}

\item{path}{To which subfolder the data is saved. Default to '../data'. If './', save to current directory.}

\item{compress}{Whether the object should be compressed}

\item{rds}{Whether the object should be saved as '.rds' file. The default value is 'TRUE', which the object will be saved as '.Rdata' file.}
}
\description{
A simplified save function using saveRDS() or save().
}
\examples{
sv(dt) # equals to sv('dt')
sv(dt, dt_new_name, path='../Rdata')
}
