% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{ld}
\alias{ld}
\title{A simplified load function.}
\usage{
ld(x, force = F)
}
\arguments{
\item{x}{The object to be loaded.}

\item{force}{Whether the object should be reloaded if it's already in the current environment.}
}
\description{
A simplified load function.
}
