% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ld}
\alias{ld}
\title{A simplified load function using readRDS(), load() or read_feather().}
\usage{
ld(filename, objname = NULL, ldtype = "Rdata", path = "./data", force = F)
}
\arguments{
\item{filename}{expression or char. The file name to be loaded, e.g. ld(x) equals to load(file='dt.rds')}

\item{objname}{expression or char. The object name in R. If NULL then equals to filename.}

\item{ldtype}{`Rdata`, `rds` or `feather`. The default is `Rdata`.}

\item{path}{From which subfolder to read the data. Default to '../data'. If './', load from current directory.}

\item{force}{Whether the object should be reloaded if it's already in the current environment.}
}
\description{
A simplified load function using readRDS(), load() or read_feather().
}
\examples{
ld(dt)
ld(filename=dt, objname=dt_new, path='../Rdata')
}
