% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{sv}
\alias{sv}
\title{A simplified save function.}
\usage{
sv(x, compress = T)
}
\arguments{
\item{x}{The object to be saved}

\item{compress}{Whether the object should be compressed}
}
\description{
A simplified save function.
}
